# Conflicts:
#	zuul-framework/src/JavaFXSimple/GUIFX.java

+++ zuul-framework/src/TWoT_test/TWoT.java
@@ -861,7 +861,7 @@ public class TWoT implements Serializable{
      * 
      * @return 
      */
-    public int getPlayerHealth(){
+    public double getPlayerHealth(){
         return player.getHealth();
     }
     /**
